backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
pgvector==0.2.5
openai>=1.50.0
pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
alembic==1.13.1
python-jose[cryptography]
passlib[bcrypt]
python-multipart
langchain-text-splitters
PyMuPDF
python-docx
